# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"OTP Authentication API","description":"Mobile-based OTP authentication system with JWT","contact":{"name":"API Support","email":"support@example.com"},"version":"1.0.0"},"servers":[{"url":"http:\/\/localhost:8000","description":"API Server"}],"paths":{"\/api\/v1\/auth\/check-mobile":{"post":{"tags":["Authentication"],"summary":"Check if mobile number exists","description":"Check whether a mobile number is registered in the system","operationId":"checkMobile","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["mobile_number"],"properties":{"mobile_number":{"description":"Mobile number (10-15 digits)","type":"string","example":"9145813194"}},"type":"object"}}}},"responses":{"200":{"description":"Mobile number check result","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Mobile number checked successfully"},"data":{"properties":{"exists":{"type":"boolean","example":true},"is_verified":{"type":"boolean","example":true}},"type":"object"}},"type":"object"}}}},"422":{"description":"Validation error","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Validation failed"}},"type":"object"}}}}}}},"\/api\/v1\/auth\/login":{"post":{"tags":["Authentication"],"summary":"Login with password","description":"Authenticate user with mobile number and password","operationId":"login","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["mobile_number","password"],"properties":{"mobile_number":{"description":"Mobile number","type":"string","example":"9145813194"},"password":{"description":"Password (min 8 characters)","type":"string","example":"MyPassword123"}},"type":"object"}}}},"responses":{"200":{"description":"Login successful","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Login successful"},"data":{"properties":{"user":{"properties":{"id":{"type":"integer","example":1},"mobile_number":{"type":"string","example":"9145813194"},"first_name":{"type":"string","example":"John"},"last_name":{"type":"string","example":"Doe"}},"type":"object"},"token":{"type":"string","example":"eyJ0eXAiOiJKV1QiLCJhbGci..."}},"type":"object"}},"type":"object"}}}},"401":{"description":"Authentication failed","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Invalid credentials"}},"type":"object"}}}},"422":{"description":"Validation error"}}}},"\/api\/v1\/auth\/request-otp":{"post":{"tags":["Authentication"],"summary":"Request OTP code","description":"Send OTP code via SMS for registration or verification","operationId":"requestOtp","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["mobile_number"],"properties":{"mobile_number":{"description":"Mobile number to receive OTP","type":"string","example":"9145813194"}},"type":"object"}}}},"responses":{"200":{"description":"OTP sent successfully","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"OTP sent successfully"},"data":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"OTP sent successfully"},"otp_code":{"description":"OTP code (only in debug mode)","type":"string","example":"123456"},"expires_at":{"type":"string","format":"date-time","example":"2025-10-07T10:00:00Z"}},"type":"object"}},"type":"object"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Rate limit exceeded"}},"type":"object"}}}},"422":{"description":"Validation error"}}}},"\/api\/v1\/auth\/register":{"post":{"tags":["Authentication"],"summary":"Complete registration with OTP","description":"Verify OTP and complete user registration","operationId":"register","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["mobile_number","otp_code","first_name","last_name","national_id","password","password_confirmation"],"properties":{"mobile_number":{"description":"Mobile number","type":"string","example":"9145813194"},"otp_code":{"description":"6-digit OTP code","type":"string","example":"123456"},"first_name":{"description":"First name","type":"string","example":"John"},"last_name":{"description":"Last name","type":"string","example":"Doe"},"national_id":{"description":"10-digit national ID","type":"string","example":"1234567890"},"password":{"description":"Password (min 8 characters)","type":"string","example":"MyPassword123"},"password_confirmation":{"description":"Password confirmation","type":"string","example":"MyPassword123"}},"type":"object"}}}},"responses":{"201":{"description":"Registration completed successfully","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Registration completed successfully"},"data":{"properties":{"user":{"properties":{"id":{"type":"integer","example":1},"mobile_number":{"type":"string","example":"9145813194"},"first_name":{"type":"string","example":"John"},"last_name":{"type":"string","example":"Doe"},"national_id":{"type":"string","example":"1234567890"},"is_mobile_verified":{"type":"boolean","example":true}},"type":"object"},"token":{"type":"string","example":"eyJ0eXAiOiJKV1QiLCJhbGci..."}},"type":"object"}},"type":"object"}}}},"400":{"description":"Invalid or expired OTP","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Invalid or expired OTP code"}},"type":"object"}}}},"422":{"description":"Validation error"}}}},"\/api\/v1\/auth\/me":{"get":{"tags":["Authentication"],"summary":"Get authenticated user","description":"Retrieve current authenticated user information","operationId":"me","responses":{"200":{"description":"User retrieved successfully","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"User retrieved successfully"},"data":{"properties":{"id":{"type":"integer","example":1},"mobile_number":{"type":"string","example":"9145813194"},"first_name":{"type":"string","example":"John"},"last_name":{"type":"string","example":"Doe"},"national_id":{"type":"string","example":"1234567890"},"is_mobile_verified":{"type":"boolean","example":true},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}},"type":"object"}},"type":"object"}}}},"401":{"description":"Unauthorized - Invalid or missing token","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Unauthorized"}},"type":"object"}}}}},"security":[{"bearer_token":[]}]}}},"components":{"securitySchemes":{"bearer_token":{"type":"http","description":"Enter JWT Bearer token","bearerFormat":"JWT","scheme":"bearer"},"api_key_security":{"type":"apiKey","description":"A short description for security scheme","name":"api_key","in":"header"}}},"tags":[{"name":"Authentication","description":"Authentication and user management endpoints"}]}